changes are going well - a few more to go
